requirements.txt
annotated-types==0.7.0
anyio==4.4.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
fsspec==2024.6.0
ftfy==6.2.0
google==3.0.0
google-cloud==0.34.0
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-tools==1.64.1
h11==0.14.0
h2==4.1.0
h5py==3.11.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
lz4==4.3.3
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
packaging==24.1
pillow==10.3.0
pinecone-client==4.1.1
pinecone-plugin-interface==0.0.7
portalocker==2.10.0
protobuf==5.27.1
protoc-gen-openapiv2==0.0.1
pydantic==2.7.4
pydantic_core==2.18.4
pywin32==306
qdrant-client==1.9.2
regex==2024.5.15
requests==2.32.3
setuptools==69.5.1
sniffio==1.3.1
soupsieve==2.5
sympy==1.12.1
tbb==2021.13.0
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.3
wheel==0.43.0
